resume_ai_agent/requirements.txt
customtkinter==5.2.2
python-dotenv==1.0.1
requests==2.32.3